+++ models/seeds.js
@@ -1,0 +1,19 @@
+var tasks = [
+  {id: 1, body: "Write final brand statement", completed: false},
+  {id: 2, body: "Complete Resume worksheet", completed: false},
+  {id: 3, body: "Write resume rough draft", completed: false},
+  {id: 4, body: "Write resume final version", completed: false},
+  {id: 5, body: "Build portfolio", completed: false}
+]
+
+module.exports = tasks
+
+// var lists = [
+//   {id: 1, body: "test final brand statement", completed: false},
+//   {id: 2, body: "Complete Resume worksheet", completed: false},
+//   {id: 3, body: "Write resume rough draft", completed: false},
+//   {id: 4, body: "Write resume final version", completed: false},
+//   {id: 5, body: "Build portfolio", completed: false}
+// ]
+//
+// module.exports = lists
