+++ views/index.hbs
@@ -1,0 +1,4 @@
+<h1>Welcome to Do Something</h1>
+
+<a href='/tasks'>See all Tasks</a><br>
+<a href='/lists'>See all Lists</a>
